requirements.txt
numpy==1.18.5
tqdm==4.61.0
matplotlib==3.3.2
mediapipe==0.8.4.2
opencv_python==4.2.0.34
Pillow==8.3.2
streamlit==0.88.0
